merge: incorporate builtin call extraction
