+++ package.json
@@ -8,6 +8,7 @@
   "exports": "./dist/index.js",
   "scripts": {
     "test": "tsc --noEmit && vitest run",
+    "test:w": "vitest",
     "snapshot": "vitest -u",
     "build": "npx tsc",
     "prepublishOnly": "tsc"
