+++ src/parser/lexer.ts
@@ -8,91 +8,115 @@ import {
   isDigitSign,
 } from "./grammar.js";
 
-export const lexer = (chars: CharStream): Token => {
-  const token = new Token({
-    location: chars.currentSourceLocation(),
-  });
+/**
+ * Lexer that tracks angle bracket nesting depth so that `>>` can be
+ * tokenized either as two closing brackets (inside generics) or as a shift
+ * operator (outside generics).
+ */
+export class Lexer {
+  private angleBracketDepth = 0;
 
-  while (chars.hasCharacters) {
-    const char = chars.next;
+  tokenize(chars: CharStream): Token {
+    const token = new Token({
+      location: chars.currentSourceLocation(),
+    });
 
-    if (!token.hasChars && char === " ") {
-      consumeSpaces(chars, token);
-      break;
-    }
+    while (chars.hasCharacters) {
+      const char = chars.next;
 
-    if (!token.hasChars && nextIsNumber(chars)) {
-      consumeNumber(chars, token);
-      break;
-    }
+      if (!token.hasChars && char === " ") {
+        this.consumeSpaces(chars, token);
+        break;
+      }
 
-    if (!token.hasChars && char === ",") {
+      if (!token.hasChars && this.nextIsNumber(chars)) {
+        this.consumeNumber(chars, token);
+        break;
+      }
+
+      if (!token.hasChars && char === ",") {
+        token.addChar(chars.consumeChar());
+        break;
+      }
+
+      if (!token.hasChars && isOpChar(char)) {
+        this.consumeOperator(chars, token);
+        break;
+      }
+
+      if (!token.hasChars && isTerminator(char)) {
+        token.addChar(chars.consumeChar());
+        break;
+      }
+
+      // Support sharp identifiers (Used by reader macros ignores non-whitespace terminators)
+      if (token.first === "#" && !isWhitespace(char)) {
+        token.addChar(chars.consumeChar());
+        continue;
+      }
+
+      if (char === "\t") {
+        throw new Error(
+          "Tabs are not supported, use four spaces for indentation",
+        );
+      }
+
+      if (isTerminator(char)) {
+        break;
+      }
+
       token.addChar(chars.consumeChar());
-      break;
     }
 
-    if (!token.hasChars && isOpChar(char)) {
-      consumeOperator(chars, token);
-      break;
-    }
+    token.setEndLocationToStartOf(chars.currentSourceLocation());
 
-    if (!token.hasChars && isTerminator(char)) {
+    this.updateAngleBracketDepth(token, chars);
+
+    return token;
+  }
+
+  private consumeOperator(chars: CharStream, token: Token) {
+    while (isOpChar(chars.next)) {
+      if (token.value === ">" && (this.angleBracketDepth > 0 || chars.next === ":")) {
+        break;
+      }
+
       token.addChar(chars.consumeChar());
-      break;
     }
+  }
 
-    // Support sharp identifiers (Used by reader macros ignores non-whitespace terminators)
-    if (token.first === "#" && !isWhitespace(char)) {
+  private updateAngleBracketDepth(token: Token, chars: CharStream) {
+    if (token.value === "<" && !isWhitespace(chars.next)) {
+      this.angleBracketDepth += 1;
+    } else if (token.value === ">" && this.angleBracketDepth > 0) {
+      this.angleBracketDepth -= 1;
+    }
+  }
+
+  private consumeNumber(chars: CharStream, token: Token) {
+    const isValidNumber = (str: string) =>
+      /^[+-]?\d+(?:\.\d+)?([Ee]?[+-]?\d+|(?:i|f)(?:|3|6|32|64))?$/.test(str);
+    const stillConsumingNumber = () =>
+      chars.next &&
+      (isValidNumber(token.value + chars.next) ||
+        isValidNumber(token.value + chars.next + chars.at(1)));
+
+    while (stillConsumingNumber()) {
       token.addChar(chars.consumeChar());
-      continue;
     }
-
-    if (char === "\t") {
-      throw new Error(
-        "Tabs are not supported, use four spaces for indentation"
-      );
-    }
-
-    if (isTerminator(char)) {
-      break;
-    }
-
-    token.addChar(chars.consumeChar());
   }
 
-  token.setEndLocationToStartOf(chars.currentSourceLocation());
-  return token;
-};
+  private nextIsNumber(chars: CharStream) {
+    return (
+      isDigit(chars.next) ||
+      (isDigitSign(chars.next) && isDigit(chars.at(1) ?? ""))
+    );
+  }
 
-const consumeOperator = (chars: CharStream, token: Token) => {
-  while (isOpChar(chars.next)) {
-    if (token.value === ">" && (chars.next === ">" || chars.next === ":")) {
-      break; // Ugly hack to support generics, means >> is not a valid operator. At least until we write a custom parser for the generics reader macro.
+  private consumeSpaces(chars: CharStream, token: Token) {
+    while (chars.next === " " && token.length < 2) {
+      token.addChar(chars.consumeChar());
     }
-
-    token.addChar(chars.consumeChar());
   }
-};
+}
 
-const consumeNumber = (chars: CharStream, token: Token) => {
-  const isValidNumber = (str: string) =>
-    /^[+-]?\d+(?:\.\d+)?([Ee]?[+-]?\d+|(?:i|f)(?:|3|6|32|64))?$/.test(str);
-  const stillConsumingNumber = () =>
-    chars.next &&
-    (isValidNumber(token.value + chars.next) ||
-      isValidNumber(token.value + chars.next + chars.at(1)));
-
-  while (stillConsumingNumber()) {
-    token.addChar(chars.consumeChar());
-  }
-};
-
-const nextIsNumber = (chars: CharStream) =>
-  isDigit(chars.next) ||
-  (isDigitSign(chars.next) && isDigit(chars.at(1) ?? ""));
-
-const consumeSpaces = (chars: CharStream, token: Token) => {
-  while (chars.next === " " && token.length < 2) {
-    token.addChar(chars.consumeChar());
-  }
-};
