+++ src/parser/parse-chars.ts
@@ -1,7 +1,7 @@
 import { Expr } from "../syntax-objects/expr.js";
 import { List } from "../syntax-objects/list.js";
 import { CharStream } from "./char-stream.js";
-import { lexer } from "./lexer.js";
+import { Lexer } from "./lexer.js";
 import { getReaderMacroForToken } from "./reader-macros/index.js";
 import { Token } from "./token.js";
 
@@ -9,25 +9,27 @@ export type ParseCharsOpts = {
   nested?: boolean;
   terminator?: string;
   parent?: Expr;
+  lexer?: Lexer;
 };
 
 export const parseChars = (
   file: CharStream,
   opts: ParseCharsOpts = {}
 ): List => {
+  const lexer = opts.lexer ?? new Lexer();
   const list = new List({
     location: file.currentSourceLocation(),
     parent: opts.parent,
   });
 
   while (file.hasCharacters) {
-    const token = lexer(file);
+    const token = lexer.tokenize(file);
 
     if (token.value === opts.terminator) {
       break;
     }
 
-    if (processWithReaderMacro(token, list.last(), file, list)) {
+    if (processWithReaderMacro(token, list.last(), file, list, lexer)) {
       continue;
     }
   }
@@ -41,7 +43,8 @@ const processWithReaderMacro = (
   token: Token,
   prev: Expr | undefined,
   file: CharStream,
-  list: List
+  list: List,
+  lexer: Lexer,
 ) => {
   const readerMacro = getReaderMacroForToken(token, prev, file.next);
   if (!readerMacro) return undefined;
@@ -52,6 +55,7 @@ const processWithReaderMacro = (
       parseChars(file, {
         nested: true,
         terminator,
+        lexer,
       }),
   });
